Snapshot all auths during a test not just last invokes (#1168)
### What
Snapshot all auths during a test not just last invokes.

### Why
The test snapshot logic is using the authorization manager of the host
to get auths that have occurred. However, the authorization manager only
has the auths from the last invoke. To get all auths for the Envs
lifetime the contract invocation hook needs to be used to tell us when
each call to the host finishes so that the auths can be collected as the
test executes.

Close #1164

### Merging
Intended for merging to `main` after:
- #1167