Improvements to Arbitrary Options (#1163)
### What

Simplify generic type bounds related to `ArbitraryOption`. Add tests.
Add `Option` to `ArbitraryVal`.

### Why

The bounds on this type are more complex than other containers like
`Vec<T>`. The reason is that the `TryFromVal` implementation assumes
that the conversions it makes are fallible and propagates errors that
necessitate complex bounds related to the error types.

These conversions are infallible though and don't need to worry about
errors.

The third commit makes the impl of `Arbitrary` for `Val` generate
`Option`s.

### Known limitations

na

---

I also slipped in a minor fix for `ArbitraryValVec::Map`, where it was
generating the wrong type.

---------

Co-authored-by: Leigh McCulloch <351529+leighmcculloch@users.noreply.github.com>