Hide EnvBase and other internal types (#1172)
### What
Hide EnvBase and other internal types in docs.

### Why
EnvBase is an implementation detail and not intended to be part of the
Soroban SDKs public interface.

The intent was for it to be hidden, but the way we tried to hide it
doesn't work for traits, even though it works for other types.

Close #1171