Restore previous auth state in client (#1174)
### What
Restore previous auth state in client after client finishes invoke.

### Why
The client is presented in such a way that it looks like invocations
with a client that has been configured with certain auth will apply the
auth only during invocations with it.

It's unintuitive that the auths from the client apply forever after the
invocation.

Close #1084

Dependent on:
- https://github.com/stellar/rs-soroban-env/pull/1244