Make try_ fns return smaller SDK error type (#1175)
### What
Make try_ fns return smaller SDK error type instead of the large host
error type.

### Why
The host error type contains a massive number of possible errors.
However, the host obfuscates the error when it is passed to the guest,
and so the guest and sdk only ever see the Context/InvalidAction error.

It's confusing for developers if we expose the host error type, then
don't actually surface most of the errors on it. A developer might think
they can distingish between different error states, when in reality they
cannot.


Close #1136