Remove unused CI step (#1207)
### What
Remove CI step that installs gcc/g++ for aarch64 on Linux.

### Why
It doesn't appear to be needed. The step has an error in it's `if`
clause that causes it to never be triggered. The if checks
`matrix.target` but should be checking `matrix.sys.target`. All the CI
tests are passing so it seems like it is not needed.