Fix bug with Timepoint/Duration as parameters/returns from functions (#1200)
### What
Map contract function parameter types `Timepoint` and `Duration` to the
respective spec types in the contract spec.

### Why
They aren't being mapped today which results in them being added to the
spec as custom types. This is a bug, they aren't custom types, and
causes generated code to incorrectly generate clients for calling the
contract.

Close #1196