Add storage update fns (#1209)
### What
Add fns to the storage interface for updating (getting then setting)
key-value pairs.

### Why
In contracts when I want to do a get and set, an interface like this
where I only need to provide the key once, and select the storage type
once, is less noisy.

Here's a small example from the `mint-lock` example in the
soroban-examples repo. It's not really much less lines of code, but line
count isn't the motivation here, rather grouping the "open" and "close"
of the update.
```diff
-            let minter_stats = env
-                .storage()
-                .temporary()
-                .get::<_, MinterStats>(&minter_stats_key)
-                .unwrap_or_default();
-            let new_minter_stats = MinterStats {
-                consumed_limit: minter_stats.consumed_limit + amount,
-            };
-            if new_minter_stats.consumed_limit > config.limit {
-                return Err(Error::DailyLimitInsufficient);
-            }
-            env.storage()
-                .temporary()
-                .set::<_, MinterStats>(&minter_stats_key, &new_minter_stats);
+            env.storage().temporary().try_update(
+                &minter_stats_key,
+                |val: Option<MinterStats>| {
+                    let mut val = val.unwrap_or_default();
+                    val.consumed_limit += amount;
+                    if val.consumed_limit > config.limit {
+                        return Err(Error::DailyLimitInsufficient);
+                    }
+                    Ok(val)
+                },
+            )?;

```