Allow &Env in contract fns (#1210)
### What
Allow the first param of contract functions to be &Env or Env, instead
of requiring that if present it is Env.

### Why
To reduce repetitive clones in contract functions.

Requiring as input to functions an owned copy of Env is unnecessary, and
it has a run on affect that if someone wants to use a contract function
also as an internal function called by their code they need to clone the
Env each time it is called. This has zero performance implications
because the Env is, or should be, erased from most places at build time,
but it adds noise to the application. Looking back we should have made
the Env a ref everywhere, probably, but in lieu of doing that which
would be a breaking change we can make it allowed so both are allowed.

This will also remove a compile fail case for devs. Before if they used
an &Env they would see an error. Now they won't.

This change is not a breaking change.