Revert "Bump rs-soroban-env dependency" (#1212)
### What

Reverts stellar/rs-soroban-sdk#1208

### Why

To release in v20.2.0 of the SDK (#1211) without updating the version of
the env.

I'll reapply the env change to the repo after the release.

Because of the number of other smaller changes in the repo it is easier
to revert this one commit and reapply it than it is to fork the main
branch for a release.