Add option to disable test snapshots on Env (#1235)
### What
Add option to disable test snapshots on Env

### Why
In some situations a developer needs to disable test snapshots. The
default is that they are always on, in the interest of giving all
developers a way to identify unexpected and otherwise untested for
changed behavior during SDK upgrades, no protocol releases, or their own
developer lifecycle.

Close #1225

### Merging
This change is dependent on the following change merging first, and will
be kept in draft state until then:
- #1234