Add doc qualifying where contractimport is relative to (#1241)
### What
Add doc qualifying where the path passed to contractimport is relative
to.

### Why
The docs don't say, and you could easily assume it is relative to the
current file which is what some built-in Rust macros do. The reason it
isn't relative to the current file is because Rust macros don't provide
anyway for a non-built-in macro to do that.