Add docs to the token interface (#1247)
### What
Add docs to the token interface, and a summary that outlines how the
different functions are grouped into areas of functionality.

### Why
To help a newcomer engaging with the interface to see which functions
are related to each other, and then when digging into the allowance
functions specifically, give more understanding and depth in how the
functions are intended to work.

Close #966