Do not generate client fns for reserved hook functions (#1277)
### What
Do not generate client fns for reserved hook functions.

### Why
There's no value in generating client functions for reserved hook
functions (that begin with '__') because the Soroban Env won't let them
be invoked directly anyway. So by generating the functions we're
creating code that can never successfully be called.

Note that the function should still be generated into the
interface/trait so that someone implementing the contracts trait
implements the entire contract and not a subset.

Related to https://github.com/stellar/stellar-cli/issues/1344