Add util to update issuer flags for SAC testing (#1320)
### What

Resolves #1117 

Instead of making a general `storage` utility to modify the account, I
added a specific method to update the issuer flag. because there isn't
any reason to mess with the other fields directly.

I didn't want to expose the xdr in the interface, but doing that
would've made the interface more clear.

---------

Co-authored-by: Leigh McCulloch <351529+leighmcculloch@users.noreply.github.com>