Fix testutils invoke with incorrect arg count (#1336)
### What

Panic if the number of args being passed to a natively registered
contract don't match the number of args the contract function expects.

### Why

This bug exists only in the testutils path of invoking a natively
registered contract and does not impact WASM registered contracts, or
the env as deployed.

If a native contract is invoked with too few arguments, a panic occurs
with an index out of bounds errors that is not intuitive to debug. The
panic displayed to the developer should be as intuitive as possible.
- Before: `index out of bounds: the len is 1 but the index is 1`
- After: `invalid number of input arguments: 2 expected, got 1`

If a native contract is invoked with too many arguments, no panic
occurs, the additional arguments are simply ignored. The behaviour of
invoking natively registered contracts in testutils should be as
consistent as possible to invoking contracts on a deployed network. In
both of these cases the invoke fails.
- Before: no panic
- After: `invalid number of input arguments: 2 expected, got 3`

Thanks to @dmkozh who fixed this bug and proposed the better error
experience in #1327 which is destined for the next major release of the
sdk v22, and this pull request extracts the fix in isolation so we can
release a bug fix release of the sdk v21.

---------

Co-authored-by: Dmytro Kozhevin <dmytro@stellar.org>