Fix warnings on contract and contractimpl macros (#1344)
### What
Fix and disable warnings on contract and contractimpl macros for
misformatted item names, and missing fns.

### Why
Starting with the last release warnings started to display on the
contract and contractimpl macros because of generated code not following
the naming conventions for some items.

In one case it is easier to fix the naming convention, and so that is
what was done.

In the other case it is easier to silence the warning, and so that is
what was done.

One of the warnings was about a missing function. In this case I think
there's an issue with rust-analyzer and how we were generating two
dependent parts of code different ways. In one are we were always
generating the code feature gated, in the other we were generating it
only based on the feature. Rust-analyzer doesn't rebuild proc-macros
frequently, and so I think caching of generated code that then had the
feature off was to blame.

### Backporting

This change should be backported to the 21 versions as a patch on 21.7
after merging to main.