Use ConstructorArgs trait for args on deploy with constructor (#1349)
### What
Use ConstructorArgs trait for args on deploy with constructor, and
rename deploy with constructor to deploy_v2.

### Why
So that types can be passed directly to the deploy function like they
can now be passed to the register function since #1343.

I renamed the function to deploy_v2 for consistency with what we did
when we added a second register for stellar asset contract function. I
considered merging with the deploy function which would have been a
breaking change, that I think we agreed would be worthwhile, but
breaking is just such a bad experience and I think a _v2 is almost as
good. We can consider merging the deploy functions at a later date.

Close #1347