Add compile errors when built with unsupported wasm features (#1353)
### What
Add compile errors when built with unsupported wasm features
reference-types and multivalue.

### Why
Rust 1.82 is likely to ship with target feature reference-types and
multivalue enabled on wasm builds. These target features are not
supported by the Soroban Env in the current protocol 21 or next planned
protocol 22. It's not trivial for developers to disable target features
because of how the rustc compiler only exposes the ability to buildstd
with nightly.

These compile errors will prevent someone from building .wasm files with
the sdk when either of those target features are enabled.

A Rust version check is not being done because it's actually easier to
do a target feature check, and in the off chain somehow Rust 1.82
shipped without the target features enabled then the sdk could still be
used with 1.82.

Links:
- https://discord.com/channels/897514728459468821/1289090985569026048
- https://github.com/stellar/rs-soroban-env/issues/1469
- https://github.com/WebAssembly/tool-conventions/issues/233


### Releasing
This change will be merged to `main` targeting v22, then should be
backported to a v21 patch release immediately following.